am243x: DPL: Fix MisraC and SA issues
Fix all the MISRA-C violations for DPL No RTOS files
for Common, m4 and r5 files

Fixes: MCUSDK-10911

+++ source/drivers/ipc_notify/v1/soc/am273x/ipc_notify_v1_cfg.c
@@ -214,7 +214,7 @@ void IpcNotify_trigInterrupt(uint32_t mailboxBaseAddr, uint32_t intrBitPos)
     volatile uint32_t *addr = (uint32_t *)mailboxBaseAddr;
 
     /* trigger interrupt to other core */
-    *addr = (1U << intrBitPos);
+    *addr = ((uint32_t)1U << intrBitPos);
 }
 
 void IpcNotify_wait(void)
